WIP: Refactoring CLI and REPL logic, snapshot of partially working state.
