+++ src/sidekick/session.py
@@ -4,7 +4,7 @@ user_config = {}
 agents = {}
 messages = []
 total_cost = 0.0
-current_model = ""  # Will be set from user_config after setup
+current_model = "openai:gpt-4o"  # Will be set from user_config after setup
 spinner = None
 tool_ignore = []  # Tools to ignore during confirmation
 yolo = False  # Skip all confirmations if true
@@ -12,3 +12,4 @@ undo_initialized = False  # Whether the undo system has been initialized
 session_id = str(uuid.uuid4())  # Unique ID for the current session
 device_id = None  # Unique ID for the device, loaded during initialization
 telemetry_enabled = True
+current_task = None
