--- src/sidekick/tools/__init__.py
@@ -1,11 +1,0 @@
-from .read_file import read_file
-from .run_command import run_command
-from .update_file import update_file
-from .write_file import write_file
-
-__all__ = [
-    "run_command",
-    "read_file",
-    "write_file",
-    "update_file",
-]
