+++ src/sidekick/tools/run_command.py
@@ -1,6 +1,12 @@
 import subprocess
 
-from sidekick.utils import ui
+
+message_handler = None
+
+
+def _msg(message_type: str, message: str):
+    if message_handler:
+        message_handler(message_type, message)
 
 
 def run_command(command: str) -> str:
@@ -14,7 +20,7 @@ def run_command(command: str) -> str:
         str: The output of the command (stdout and stderr) or an error message.
     """
     try:
-        ui.status(f"Shell({command})")
+        _msg("status", f"Shell({command})")
 
         process = subprocess.Popen(
             command,
@@ -31,7 +37,7 @@ def run_command(command: str) -> str:
         # Raise retry if the output is too long to prevent issues
         # Reduced limit as it's often better to be concise
         if len(resp) > 5000:
-            ui.warning("Command output too long, returning truncated.")
+            # _msg("warning", "Command output too long, returning truncated.")
             # Include both the beginning and end of the output
             start_part = resp[:2500]
             end_part = resp[-1000:] if len(resp) > 3500 else resp[2500:]
@@ -40,11 +46,10 @@ def run_command(command: str) -> str:
 
         return resp
     except FileNotFoundError as e:
-        # Specific error for command not found
         err_msg = f"Error: Command not found or failed to execute: {command}. Details: {e}"
-        ui.error(err_msg)
+        # _msg("error", err_msg)
         return err_msg
-    except Exception as e:
-        err_msg = f"Error running command '{command}': {e}"
-        ui.error(err_msg)
-        return err_msg
+    # except Exception as e:
+    #     err_msg = f"Error running command '{command}': {e}"
+    #     _msg("error", err_msg)
+    #     return err_msg
